More work on memory subsystem - now decoupled from CPU

+++ memorymap.h
@@ -2,9 +2,9 @@
 #define MEMORYMAP_H
 int memorymap_initialise(void);
 uint32_t memorymap_read(uint32_t address, uint8_t width, uint32_t *value);
-int  memorymap_write(uint32_t address, uint8_t width, uint32_t value);
+int  memorymap_write(uint32_t address, uint32_t width, uint32_t value);
 int  memorymap_aligned_read(uint32_t address, uint32_t *value);
-int  memorymap_aligned_write(uint32_t address, uint8_t mask, uint32_t value);
+int  memorymap_aligned_write(uint32_t address, uint32_t mask, uint32_t value);
 void memorymap_dump(void);
 void memorymap_dump(void);
 void memorymap_finish(void);
